refactor(gax-internal): Organize observability into sub-modules\n\nRestructures the observability.rs file into a directory module\nwith sub-modules for attributes and http_tracing to improve\norganization. No functional changes.
