+++ src/gax-internal/src/lib.rs
@@ -36,7 +36,7 @@ pub mod query_parameter;
 #[cfg(feature = "_internal-http-client")]
 pub mod http;
 
-#[cfg(feature = "_internal-http-client")]
+#[cfg(all(feature = "_internal-http-client", feature = "_unstable-o12y"))]
 pub mod observability;
 
 #[cfg(feature = "_internal-grpc-client")]
